% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_gmed.R
\name{bayes_gmed}
\alias{bayes_gmed}
\title{Estimate a causal mediation model}
\usage{
bayes_gmed(
  outcome,
  mediator,
  treat,
  covariates = NULL,
  dist.y = "continuous",
  dist.m = "continuous",
  link.y = "identity",
  link.m = "identity",
  data,
  control.value = 0,
  treat.value = 1,
  priors = NULL,
  ...
)
}
\arguments{
\item{outcome}{a character string indicating the name of the outcome variable.}

\item{mediator}{a character vector indicating the name of the confounding variables.}

\item{treat}{a character string indicating the name of the treatment variable. The treatment can be either binary (integer or a two-valued factor) or continuous (numeric).}

\item{dist.y}{a character string indicating the family distribution of the outcome.  E.g., dist.y = "bernoulli" will fit
a logistic regression the outcome.}

\item{dist.m}{a character string indicating the family distribution of the mediator  E.g., dist.m = "bernoulli" will fit
a logistic regression the mediator}

\item{link.y}{a character string indicating the link function to be used for the outcome model.}

\item{link.m}{a character string indicating the link function to be used for the mediator model.}

\item{data}{A \code{data.frame} or a \code{tibble} object.}

\item{control.value}{The vaue of the treatment variable to be used as a reference level.}

\item{treat.value}{The vaue of the treatment variable to be used as a treatment level.}

\item{priors}{A list of named values to be used as the prior scale parameters. See details.}

\item{...}{Other optional parameters passed to \code{rstan::stan()}.}
}
\value{
An object of S4 class stanfit, with all its available methods.
}
\description{
Estimates a Bayesian causal mediation model using g-formula approach though Stan.
}
\details{
Draw samples from the joint posterior distribution of a
mediation model using Stan.
}
\author{
Belay Birlie Yimer \email{belaybirlie.yimer@manchester.ac.uk}
}
