% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{example_data: A simulated data set for illustrating mediation analysis in BayesGmed.}
\format{
\subsection{\code{example_data}}{

A data frame with 300 rows and 5 columns:
\describe{
\item{Z1, Z2}{confounders}
\item{A}{binary exposure}
\item{M}{binary mediator}
\item{Y}{binary outcome}
...
}
}
}
\usage{
example_data
}
\description{
The data set contains a binary outcome variable Y, a binary exposure A,
a binary mediator M, and two confounding variables (Z1 and Z2) for
300 individuals.
}
\keyword{datasets}
