% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesgmed_sens.R
\name{bayesgmed_sens}
\alias{bayesgmed_sens}
\title{bayesgmed_sens: Conduct sensitivity analysis for unmeasured confounder in causal mediation analysis}
\usage{
bayesgmed_sens(
  outcome,
  mediator,
  treat,
  covariates = NULL,
  dist.y = "continuous",
  dist.m = "continuous",
  link.y = "identity",
  link.m = "identity",
  data,
  control.value = 0,
  treat.value = 1,
  priors = NULL,
  ...
)
}
\arguments{
\item{outcome}{a character string indicating the name of the outcome variable.}

\item{mediator}{a character string indicating the name of the mediator variable.}

\item{treat}{a character string indicating the name of the treatment variable. The treatment can be either binary (integer or a two-valued factor) or continuous (numeric).}

\item{covariates}{a character vector indicating the name of the confounding variables.}

\item{dist.y}{a character string indicating the family distribution of the outcome.  E.g., dist.y = "bernoulli" will fit
a logistic regression the outcome.}

\item{dist.m}{a character string indicating the family distribution of the mediator  E.g., dist.m = "bernoulli" will fit
a logistic regression the mediator}

\item{link.y}{a character string indicating the link function to be used for the outcome model.}

\item{link.m}{a character string indicating the link function to be used for the mediator model.}

\item{data}{A \code{data.frame} or a \code{tibble} object.}

\item{control.value}{The vaue of the treatment variable to be used as a reference level.}

\item{treat.value}{The vaue of the treatment variable to be used as a treatment level.}

\item{priors}{A list of named values to be used as the prior scale parameters. See details.}

\item{...}{Other optional parameters passed to \code{rstan::stan()}.}
}
\value{
An object of S4 class stanfit, with all its available methods.
}
\description{
Estimates a Bayesian causal mediation model using g-formula approach though Stan.
}
\details{
Bayesgmed_sens allows one to conduct a sensetivity analysis for unmeasured confounding according
to the approch proposed by McCandless LC and Somers JM (2019). This is done by by incorporating uncertainty
about unmeasured confounding in the outcome and mediator model through a prior distribution. One can control
the size of unmeasured confounding by varying the values of the location and scale parameters of the prior distribution for the bias (i.e., gamma).
\subsection{priors}{

Users may pass a list of named values for the priors argument. The values will be used to define
the scale parameter of the respective prior distributions. This list may specify some or all of the
following parameters:
priors <- list(
scale_m = 2.5, scale_y = 2.5,
location_m = 0, location_y = 0,
location_gamma = 0.5, scale_gamma = 0.5,
scale_sd_y = 2.5, scale_sd_m = 2.5)
}
}
\author{
Belay Birlie Yimer \email{belaybirlie.yimer@manchester.ac.uk}
}
